Feat: The agent operator and message operator can only select string variables as prompt words. #10427
