+++ web/src/pages/agent/form/message-form/index.tsx
@@ -14,6 +14,7 @@ import { memo } from 'react';
 import { useFieldArray, useForm } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
 import { z } from 'zod';
+import { JsonSchemaDataType } from '../../constant';
 import { INextOperatorForm } from '../../interface';
 import { FormWrapper } from '../components/form-wrapper';
 import { PromptEditor } from '../components/prompt-editor';
@@ -66,6 +67,7 @@ function MessageForm({ node }: INextOperatorForm) {
                           <PromptEditor
                             {...field}
                             placeholder={t('flow.messagePlaceholder')}
+                            types={[JsonSchemaDataType.String]}
                           ></PromptEditor>
                         </FormControl>
                       </FormItem>
